+++ .ScratchPad/PersistDataTest/src/startup.ts
@@ -1,45 +1,16 @@
 import { initConfigDb } from "./data/configurationData";
 import { initPokemonDb } from "./data/pokemonData";
-import { configApiPing, updateConfiguration } from "./repositories/configurationRepository";
+import { getUpdatedAppConfiguration, configApiPing, updateConfiguration } from "./repositories/configurationRepository";
 import { pokeApiPing } from "./repositories/pokeApiRepository";
 import { loadPokemonData } from "./repositories/pokemonRepository";
 import { ConfigurationData } from "./types/configurationData";
 
-// Just a week for now
-const placeHolderStaleByDate = new Date(
-    new Date().getTime() + 7 * 24 * 60 * 60 * 1000
-).toISOString();
-
-// Call out for new configuration
-const getUpdatedAppConfiguration = async () => {
-    // Pretend API call to a "home" server that holds app config data
-    // (): result => {}
-
-    const simulatedResult = {
-        supported_generations: [
-            {
-                id: 1,
-                generation_name: "generation1",
-                description: "Red, Green, Blue, and Yellow.",
-                starting_dex_no: 1,
-                count: 3, //151
-                stale_by_dts: placeHolderStaleByDate,
-                last_modified_dts: "2025-05-08T22:04:23.244Z", // new Date().toISOString()
-            }
-            // ,{
-            //     id: 2,
-            //     generation_name: 'generation2',
-            //     description: 'Gold, Silver, and Crystal.',
-            //     starting_dex_no: 152,
-            //     count: 3, // 100
-            //     stale_by_dts: placeHolderStaleByDate,
-            //     last_modified_dts: '2025-05-08T22:01:16.292Z' // new Date().toISOString()
-            // }
-        ],
-    };
-
-    return simulatedResult;
-};
+const initOfflinePlaceholderData = () => {
+    /*
+    In case the user is offline on the first spin-up of the application,
+    it would be a good idea to put a few missingno or starters in there.
+    */ 
+}
 
 export const runStartUp = async (dataSource: string, forceUpdate: boolean) => {
     initConfigDb(dataSource);
@@ -53,4 +24,9 @@ export const runStartUp = async (dataSource: string, forceUpdate: boolean) => {
     if (pokeApiPing()) {
         loadPokemonData( forceUpdate );
     }
+
+    // Run anything needed in case user is offline
+    // - Put placeholder data in?
+
+    // Tell system that start up is complete
 };
